server: Remove most of V2 API
